+++ src/zmq.cpp
@@ -262,7 +262,7 @@ void *zmq_init (int app_threads_, int io_threads_, int flags_)
 
     //  Create 0MQ context.
     zmq::dispatcher_t *dispatcher = new (std::nothrow) zmq::dispatcher_t (
-        app_threads_, io_threads_);
+        (uint32_t) app_threads_, (uint32_t) io_threads_);
     zmq_assert (dispatcher);
     return (void*) dispatcher;
 }
